+++ projects/ngx-store/src/public_api.ts
@@ -1,1 +1,2 @@
-export { ReactiveStoreModule, StoreModule, STORE_MIDDLEWARE } from './lib/ngx-store.module';
+export * from './lib/interfaces';
+export * from './lib/store';